requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
sqlalchemy
pymysql
python-dotenv